Corrección en código de barras
Cambio de tipo de dato de Integer a Long

+++ SistemaGC/src/cr/ac/uia/SistemaGC/gui/Personas_Principal.java
@@ -219,7 +219,7 @@ public class Personas_Principal extends SGCForm {
                     "Error",
                     JOptionPane.ERROR_MESSAGE);
         } else {
-            new BarCode_Formulario((Integer) tblPersonas.getValueAt(i, 0)).setVisible(true);
+            new BarCode_Formulario((Long) tblPersonas.getValueAt(i, 0)).setVisible(true);
         }
     }//GEN-LAST:event_btnImprimirCodigoActionPerformed
